Rename tokenRegistrySubmitter to tokenRegistryProxy
