+++ server/tokenRegistryProxy.js
